data-app/professor_reviews/requirements.txt
pymongo
qdrant-client
scikit-learn
numpy
python-dotenv
ollama
tqdm